Changed parser to support 64-bit unsigned integers

+++ src/api/yajl_parse.h
@@ -35,7 +35,7 @@
  * Interface to YAJL's JSON parsing facilities.
  */
 
-#include <yajl/yajl_common.h>
+#include "yajl_common.h"
 
 #ifndef __YAJL_PARSE_H__
 #define __YAJL_PARSE_H__
@@ -88,7 +88,7 @@ extern "C" {
     typedef struct {
         int (* yajl_null)(void * ctx);
         int (* yajl_boolean)(void * ctx, int boolVal);
-        int (* yajl_integer)(void * ctx, long integerVal);
+        int (* yajl_integer)(void * ctx, long long integerVal);
         int (* yajl_double)(void * ctx, double doubleVal);
         /** A callback which passes the string representation of the number
          *  back to the client.  Will be used for all numbers when present */
